refactor(server): Split main init function with dedicated route for landing page
