+++ server/app/nodemon.json
@@ -1,9 +1,9 @@
 {
   "exec": "npm start",
   "ignore": [".git", "node_modules/*/*"],
-  "watch": [".env", "src"],
+  "watch": [".env", "./"],
   "env": {
     "NODE_ENV": "development"
   },
-  "ext": "js,graphql,xlsx"
+  "ext": "js,graphql"
 }
